# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/payments/methods": {
      "get": {
        "operationId": "PaymentsController_methods",
        "summary": "[CUSTOMER APP] List safe references to owned saved payment methods",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentMethodResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Customer App - Payments"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      }
    },
    "/api/v1/payments/methods/{id}/default": {
      "patch": {
        "operationId": "PaymentsController_makeMethodDefault",
        "summary": "[CUSTOMER APP] Make an owned saved payment method the default",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentMethodResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Customer App - Payments"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      }
    },
    "/api/v1/payments/methods/{id}": {
      "delete": {
        "operationId": "PaymentsController_archiveMethod",
        "summary": "[CUSTOMER APP] Archive an owned saved payment method",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Payment method archived"
          }
        },
        "tags": [
          "Customer App - Payments"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      }
    },
    "/api/v1/payments/checkouts": {
      "post": {
        "operationId": "PaymentsController_createCheckout",
        "summary": "[CUSTOMER APP] Create or reuse an online SumUp checkout for an owned order",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentIntentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentCheckoutResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Paid order or idempotency conflict"
          },
          "502": {
            "description": "Payment provider rejected the request"
          },
          "503": {
            "description": "Payment provider unavailable"
          }
        },
        "tags": [
          "Customer App - Payments"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      }
    },
    "/api/v1/payments/orders/{orderId}/status": {
      "get": {
        "operationId": "PaymentsController_status",
        "summary": "[CUSTOMER APP] Get an owned order payment status and optionally refresh SumUp",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refresh",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentCheckoutResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Customer App - Payments"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      }
    },
    "/api/v1/payments/webhooks/sumup": {
      "post": {
        "operationId": "PaymentsController_webhook",
        "x-audience": "Public / Customer App",
        "summary": "[PUBLIC / CUSTOMER APP] Receive SumUp notifications; state is verified through SumUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SumUpWebhookDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Notification accepted"
          }
        },
        "tags": [
          "Customer App - Payments"
        ],
        "security": [],
        "description": "Audience: Public / Customer App",
        "x-audience-label": "Public / Customer App"
      }
    },
    "/api/v1/payments/orders/{id}/collect": {
      "post": {
        "operationId": "PaymentsController_collect",
        "x-required-roles": [
          "admin",
          "employee"
        ],
        "summary": "[ADMIN APP / EMPLOYEE / DRIVER APP] Record cash or external-terminal collection on delivery",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectPaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentCheckoutResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Staff Apps - Payments"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Admin App / Employee / Driver App",
        "x-audience-label": "Admin App / Employee / Driver App"
      }
    },
...
  },
  "info": {
    "title": "La Favola Pizza Restaurant API",
    "description": "Routes are grouped by audience. Customer app developers should use sections prefixed \"Customer App\" and \"Audience: Public / Customer App\". Administrative and staff operations declare their required roles in their Audience tag and x-required-roles metadata.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CollectPaymentDto": {
        "type": "object",
        "properties": {}
      },
...
      "PaymentMethodResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "paymentMethodType": {
            "type": "string"
          },
          "cardBrand": {
            "type": "string"
          },
          "cardLast4": {
            "type": "string"
          },
          "expMonth": {
            "type": "number"
          },
          "expYear": {
            "type": "number"
          },
          "label": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "paymentMethodType",
          "isDefault",
          "createdAt"
        ]
      },
      "CreatePaymentIntentDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "idempotencyKey": {
            "type": "string"
          }
        },
        "required": [
          "orderId",
          "idempotencyKey"
        ]
      },
      "PaymentCheckoutResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "example": "pending"
          },
          "amountMinor": {
            "type": "number",
            "example": 1299
          },
          "currency": {
            "type": "string",
            "example": "EUR"
          },
          "checkoutId": {
            "type": "string"
          },
          "hostedCheckoutUrl": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "orderId",
          "status",
          "amountMinor",
          "currency"
        ]
      },
      "SumUpWebhookDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  },
  "security": [
    {
      "JWT-auth": []
    }
  ]
}